Refactor state restoration to windowsStore #4
